chk_r2: Add get-some-free discounting
Get One (or Some) free special offers are computed before multibuys as
they represent greater dicounts. This is I think is an assumption, and
cannot be applied generally. An alternative complex solution may then to
combinatorically applying discounts then choosing the cheapest total
basket cost; but I think that is out of scope of the client
specification.